+++ docs/release-notes.md
@@ -10,6 +10,7 @@ Changes:
 Fixes:
 * ([#87](https://github.com/cmln/chameleon/issues/87)) Show formedit button on the navbar
 * ([#89](https://github.com/cmln/chameleon/issues/89)) Show default icon for unknown page tools on the navbar
+* ([#91](https://github.com/cmln/chameleon/issues/91)) Ensure mediawiki.skinning.content module is loadable
 
 
 ### Chameleon 2.0